extended -report flag to support json format
