+++ Code/Tools/FBuild/FBuildCore/Helpers/FBuildStats.cpp
@@ -77,9 +77,12 @@ void FBuildStats::OnBuildStop( Node * node )
         // detailed build report
         if ( generateReport )
         {
-            Report r;
-            r.Generate( *this );
-            r.Save();
+            Report* report = options.GetReport();
+
+            report->Generate( *this );
+            report->Save();
+
+            FREE(report);
         }
 
         // stdout summary
