+++ Code/Tools/FBuild/FBuildCore/Helpers/Report.h
@@ -1,4 +1,4 @@
-// Report - Build Report Generator
+// Report - An interface for report generation
 //------------------------------------------------------------------------------
 #pragma once
 
@@ -20,43 +20,13 @@ class Node;
 class Report
 {
 public:
-    Report();
-    ~Report();
+    Report( size_t initialCapacity, bool resizeable );
+    virtual ~Report();
 
-    void Generate( const FBuildStats & stats );
-    void Save() const;
-
-private:
-    // Report sections
-    void CreateHeader();
-    void CreateTitle();
-    void CreateOverview( const FBuildStats & stats );
-    void DoCacheStats( const FBuildStats & stats );
-    void DoCPUTimeByType( const FBuildStats & stats );
-    void DoCPUTimeByItem( const FBuildStats & stats );
-    void DoCPUTimeByLibrary();
-    void DoIncludes();
-
-    void CreateFooter();
-
-    struct PieItem
-    {
-        PieItem( const char * l, float v, uint32_t c, void * u = nullptr )
-            : label( l )
-            , value( v )
-            , color( c )
-            , userData( u )
-        {
-        }
-
-        const char *    label;
-        float           value;
-        uint32_t        color;
-        void *          userData;
-
-        bool operator < ( const PieItem & other ) const { return value > other.value; }
-    };
-
+    virtual void Generate( const FBuildStats& stats ) = 0;
+    virtual void Save() const = 0;
+    
+protected:
     struct LibraryStats
     {
         const Node *    library;
@@ -70,16 +40,16 @@ private:
 
         bool operator < ( const LibraryStats & other ) const { return cpuTimeMS > other.cpuTimeMS; }
     };
-
+    
     struct IncludeStats
     {
-        const Node *    node;
+        const Node*     node;
         uint32_t        count;
         bool            inPCH;
 
-        bool operator < ( const IncludeStats & other ) const { return count > other.count; }
+        bool operator < ( const IncludeStats& other ) const { return count > other.count; }
 
-        IncludeStats *  m_Next; // in-place hash map chain
+        IncludeStats* m_Next; // in-place hash map chain
     };
 
     class IncludeStatsMap
@@ -88,37 +58,27 @@ private:
         IncludeStatsMap();
         ~IncludeStatsMap();
 
-        IncludeStats * Find( const Node * node ) const;
-        IncludeStats * Insert( const Node * node );
+        IncludeStats* Find( const Node* node ) const;
+        IncludeStats* Insert( const Node* node );
 
-        void Flatten( Array< const IncludeStats * > & stats ) const;
+        void Flatten( Array< const IncludeStats* >& stats ) const;
     protected:
-        IncludeStats * m_Table[ 65536 ];
+        IncludeStats* m_Table[65536];
         MemPoolBlock m_Pool;
     };
 
-    enum { DEFAULT_TABLE_WIDTH = 990 };
-
-    // Helpers
-    void DoTableStart( uint32_t width = DEFAULT_TABLE_WIDTH, const char * id = nullptr, bool hidden = false );
-    void DoTableStop();
-    void DoToggleSection( size_t numMore = 0 );
-    void DoSectionTitle( const char * sectionName, const char * sectionId );
-    void DoPieChart( const Array< PieItem > & items, const char * units );
-
     // Helper to format some text
-    void Write( MSVC_SAL_PRINTF const char * fmtString, ... ) FORMAT_STRING( 2, 3 );
+    void Write(MSVC_SAL_PRINTF const char* fmtString, ...) FORMAT_STRING(2, 3);
 
     // gather stats
     void GetLibraryStats( const FBuildStats & stats );
     void GetLibraryStatsRecurse( Array< LibraryStats * > & libStats, const Node * node, LibraryStats * currentLib ) const;
     void GetLibraryStatsRecurse( Array< LibraryStats * > & libStats, const Dependencies & dependencies, LibraryStats * currentLib ) const;
-    void GetIncludeFilesRecurse( IncludeStatsMap & incStats, const Node * node) const;
-    void AddInclude( IncludeStatsMap & incStats, const Node * node, const Node * parentNode) const;
+    void GetIncludeFilesRecurse( IncludeStatsMap& incStats, const Node* node ) const;
+    void AddInclude( IncludeStatsMap& incStats, const Node* node, const Node* parentNode ) const;
 
     // intermediate collected data
     Array< LibraryStats * > m_LibraryStats;
-    uint32_t m_NumPieCharts;
 
     // final output
     AString m_Output;
